Move Token class to lib/helper
