+++ src/lib/helper/index.js
@@ -1,3 +1,4 @@
 import APYCalculator from "./apy-calculator"
+import Token from "./token"
 
-export { APYCalculator }
+export { APYCalculator, Token }
